requirements.txt
flask
psycopg2-binary
flask-restx
python-dotenv
